requirements.txt
# Core API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Anthropic LLM
anthropic==0.40.0

# ML / NLP
sentence-transformers==3.3.1
faiss-cpu==1.9.0
scikit-learn==1.5.2
numpy==1.26.4

# File parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Config & validation
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
